Revert "Refactoring: Render -> Rendering"
This reverts commit a0bf09b27e1e9c068645c5b1632aa10f0516eb80.
