+++ src/main/java/com/github/wolfshotz/wyrmroost/client/render/DebugRendering.java
@@ -1,4 +1,4 @@
-package com.github.wolfshotz.wyrmroost.client.renderer;
+package com.github.wolfshotz.wyrmroost.client.render;
 
 /*import com.github.wolfshotz.wyrmroost.Wyrmroost;
 import com.mojang.blaze3d.vertex.PoseStack;
