+++ src/main/java/com/github/wolfshotz/wyrmroost/client/render/RenderHelper.java
@@ -1,4 +1,4 @@
-package com.github.wolfshotz.wyrmroost.client.renderer;
+package com.github.wolfshotz.wyrmroost.client.render;
 
 import com.github.wolfshotz.wyrmroost.WRConfig;
 import com.github.wolfshotz.wyrmroost.Wyrmroost;
