+++ src/main/java/com/github/wolfshotz/wyrmroost/client/render/entity/DragonEggRenderer.java
@@ -1,4 +1,4 @@
-package com.github.wolfshotz.wyrmroost.client.renderer.entity;
+package com.github.wolfshotz.wyrmroost.client.render.entity;
 
 import com.github.wolfshotz.wyrmroost.Wyrmroost;
 import com.github.wolfshotz.wyrmroost.entities.dragonegg.DragonEggEntity;
@@ -8,6 +8,7 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.client.model.EntityModel;
 import net.minecraft.client.model.geom.ModelPart;
 import net.minecraft.client.renderer.MultiBufferSource;
+import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.entity.EntityRenderer;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.client.renderer.texture.OverlayTexture;
