+++ src/main/java/com/github/wolfshotz/wyrmroost/client/render/entity/EmptyRenderer.java
@@ -1,6 +1,7 @@
-package com.github.wolfshotz.wyrmroost.client.renderer.entity;
+package com.github.wolfshotz.wyrmroost.client.render.entity;
 
 import net.minecraft.client.renderer.entity.EntityRenderer;
+import net.minecraft.client.renderer.entity.EntityRenderDispatcher;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.Entity;
