+++ src/main/java/com/github/wolfshotz/wyrmroost/client/render/entity/dragon/RoostStalkerRenderer2.java
@@ -1,4 +1,4 @@
-package com.github.wolfshotz.wyrmroost.client.renderer.entity.dragon;
+package com.github.wolfshotz.wyrmroost.client.render.entity.dragon;
 
 import com.github.wolfshotz.wyrmroost.client.model.entity.GeckoRoostStalkerModel;
 import com.github.wolfshotz.wyrmroost.client.model.entity.GeckoRoostStalkerModel.RoostStalkerEyesLayer;
