+++ src/main/java/com/github/wolfshotz/wyrmroost/client/render/entity/dragon/RoyalRedRenderer.java
@@ -1,6 +1,9 @@
-package com.github.wolfshotz.wyrmroost.client.renderer.entity.dragon;
+package com.github.wolfshotz.wyrmroost.client.render.entity.dragon;
 
+import com.github.wolfshotz.wyrmroost.client.model.entity.GeckoRoostStalkerModel;
+import com.github.wolfshotz.wyrmroost.client.model.entity.GeckoRoostStalkerModel.RoostStalkerEyesLayer;
 import com.github.wolfshotz.wyrmroost.client.model.entity.GeckoRoyalRedModel;
+import com.github.wolfshotz.wyrmroost.entities.dragon.RoostStalkerEntity;
 import com.github.wolfshotz.wyrmroost.entities.dragon.RoyalRedEntity;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import software.bernie.geckolib3.renderers.geo.GeoEntityRenderer;
