+++ src/main/java/com/github/wolfshotz/wyrmroost/client/render/entity/dragon/SilverGliderRenderer.java
@@ -1,6 +1,7 @@
-package com.github.wolfshotz.wyrmroost.client.renderer.entity.dragon;
+package com.github.wolfshotz.wyrmroost.client.render.entity.dragon;
 
 import com.github.wolfshotz.wyrmroost.client.model.entity.GeckoSilverGliderModel;
+import com.github.wolfshotz.wyrmroost.entities.dragon.CanariWyvernEntity;
 import com.github.wolfshotz.wyrmroost.entities.dragon.SilverGliderEntity;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import software.bernie.geckolib3.renderers.geo.GeoEntityRenderer;
